==> Rmd Files/16. age_gen_eth_bydis.Rmd <==
---
title: "age_gender_eth_bydis"
author: "Daniel Zoleikhaeian"
date: "2023-03-15"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Loading libraries
```{r}
library(plyr)
library(dplyr)
library(readxl)
library(ggplot2)
```

## Working directory
```{r}
setwd("C:/Users/danie/Documents/Joshi Lab Materials/3 Studies Dataset/Dataset Merge/012623")
```

## Total Dataset
```{r}
total_df <-read.csv('bgc_merge_cDiag123_new.csv')
```

## Age, Gender, and Latino Races Visualizations

### Age visualization
```{r}
total_df$cStudy <- factor(total_df$cStudy)
total_df$cDiagnosis4 <- factor(total_df$cDiagnosis4)

# remove the unreasonable or NA age observations
red_df <- total_df[!is.na(total_df$cAge), ]
age_vplot <- ggplot(red_df[red_df$cAge < 200, ], aes(x = cDiagnosis4, y = cAge, fill = cDiagnosis4)) + 
  geom_violin() +
  ggtitle('Age Distribution across Conditions') +
  theme(plot.title = element_text(hjust = 0.5))
  
age_plot <- age_vplot + geom_boxplot(width = 0.1)
age_plot
```

### Gender proportions by disease
```{r}
# Group by study and gender to get total counts
dis_counts <- as.data.frame(total_df %>% 
  dplyr::group_by(cDiagnosis4) %>%  
  dplyr::summarise(total_count=n(), .groups = 'drop'))

by_dis_gender <- total_df %>% 
  dplyr::group_by(cDiagnosis4, cGender) %>%  
  dplyr::summarise(total_count=n(), .groups = 'drop')

rel_freqs <- rep(0, nrow(by_dis_gender))
rel_freqs_gen <- rel_freqs

# generating relative frequencies from counts
for (i in 1:nrow(dis_counts)) {
  ind_to_divide <- by_dis_gender[, 1] == as.character(dis_counts[i, 1])
  rel_freqs_gen[ind_to_divide] <-  unname(unlist(by_dis_gender[ind_to_divide, 3] / dis_counts[i,2]))
  
}

# adding new column
by_dis_gender$Proportion<- rel_freqs_gen

gen_within <- ggplot(data = by_dis_gender, aes(x = cGender, y = Proportion, fill = cDiagnosis4)) +
  geom_bar(position = position_dodge(), 
           stat = 'identity') +
  ggtitle('Gender Proportions across Conditions') + 
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5))

gen_within
```

### Race breakdown among Latinos
```{r}
lat_df <- total_df[total_df$cHispanicorLatino == 'Yes', ]
race_cts <- plyr::count(lat_df, vars = 'cRace')

race_cts$freq <- race_cts$freq / nrow(lat_df)
colnames(race_cts)[2] <- 'Proportion'
race_cts

lat_races <- ggplot(data = race_cts, aes(x = cRace, y = Proportion, fill = cRace)) +
  geom_bar(position = position_dodge(), 
           stat = 'identity') +
  ggtitle('Race Proportions among Hispanics and Latinos') + 
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5))

lat_races

```

### Plot collection for Age, Gender, and Ethnicity
```{r}
age_plot
gen_within
lat_races
```


## Determining if differences in LR are due to Study

### Dependencies: data frames
```{r}
# Changing La Jolla to San Diego
total_df$cLocationCity[total_df$cLocationCity == 'La Jolla'] <- 'San Diego'
# Changing Brooklyn to New York
total_df$cLocationCity[total_df$cLocationCity == 'Brooklyn'] <- 'New York'

# Removing USC/MASS
total_df <- total_df[total_df$cLocationCity != 'USC/MASS', ]

df_check_2010 <- read_xlsx('reproj_30_2010.xlsx')
df_check_2020 <- read_xlsx('reproj_30_2020.xlsx')

dem_2010 <- df_check_2010[,-1]
dem_2020 <- df_check_2020[,-1]

df_avg <- df_check_2010
df_avg[,] <- NA
df_avg$City <- NULL

for (i in 1:ncol(dem_2010)) {
  df_avg[,i] <- rowMeans(cbind(dem_2010[,i], dem_2020[,i]))
  
}

df_avg <- cbind(df_check_2010$City, df_avg)
colnames(df_avg)[1] <- 'City'


# Reformatting cities
for (i in 1:length(df_avg$City)) {
  df_avg$City[i] <- substr(df_avg$City[i], 1, unlist(gregexpr(',', df_avg$City[i]))[1] - 1)
}

df_avg$City[df_avg$City == 'Stonybrook'] <- 'Stony Brook'

# Standardizing race acronyms
colnames(df_avg)[colnames(df_avg) == 'NA'] <- 'AE'
colnames(df_avg)[colnames(df_avg) == 'NH'] <- 'Not Hispanic'
colnames(df_avg)[colnames(df_avg) == 'PI'] <- 'NH'

#View(df_avg)

# Looking only at SZSAFD
szsafd <- total_df[total_df$cDiagnosis4 == 'SZSAFD', ]
cs <- total_df[total_df$cDiagnosis4 == 'CS', ]
bad12 <- total_df[total_df$cDiagnosis4 == 'BAD12',]
mdd <- total_df[total_df$cDiagnosis4 == 'MDD', ]
```

### Dependencies: Functions
```{r}
race_cities_plot_data2 <- function(null_df, tot_df, disease, study, race_acr) {
  
  obs_df <- tot_df[tot_df$cStudy %in% study & tot_df$cDiagnosis4 == disease, ]
  
  obs_cities <- obs_df$cLocationCity[obs_df$cRace == race_acr]
  
  test_cities <- intersect(null_df$City, obs_cities)
  
  null_race_cts <- null_df[, 1:8]

  null_race_props <- as.data.frame(apply(null_race_cts[,-1], MARGIN = 2, function(x) {x / sum(x)}))
  
  df_city <- data.frame(City = null_race_cts$City)
  
  null_race_props <- cbind(df_city, null_race_props)
  null_race_props <- null_race_props[null_race_props$City %in% test_cities,]
  null_race_props <- null_race_props[order(null_race_props$City), ]
  
  brc <- obs_df %>%  
    dplyr::group_by(cRace, cLocationCity) %>%  
    dplyr::summarise(total_count=n(), .groups = 'drop')
  
  brc <- brc[brc$cLocationCity %in% test_cities, ]
  
  brc <- brc[order(brc$cLocationCity), ]
  brc$Prop <- NA
  
  # Turning counts into proportions
  for (i in 1:length(test_cities)) {
    brc$Prop[brc$cLocationCity == test_cities[i]] <- brc$total_count[brc$cLocationCity == test_cities[i]] / sum(brc$total_count[brc$cLocationCity == test_cities[i]])
  }
  
  null_row <- as.data.frame(null_race_props[,c('City', race_acr)])
  obs_races <- brc[brc$cRace == race_acr, ]
  obs_prop <- obs_races$Prop
  
  ## Begin output variables to send to lm() and plot()
  propdif <- obs_prop - null_row[,2] ## response
  city_cts <- plyr::count(obs_df$cLocationCity)
  ct_to_plot <- city_cts$freq[city_cts$x %in% test_cities] ## predictor 1
  
  null_race_cts <- null_race_cts[null_race_cts$City %in% test_cities, ]
  race_of_interest_null <- null_race_cts[, colnames(null_race_cts) == race_acr] ## predictor 2
  
  null_race_sums <- as.data.frame(apply(null_race_cts[,-1], 1, sum)) ## predictor 3
  colnames(null_race_sums) <- 'Total City Population'
  
  output <- list(propdif, ct_to_plot, race_of_interest_null, null_race_sums[,1])
  names(output) <- c('Prop dif', 'city sample', 'census average race count', 'census average pop total')
  
  return (output)
  
}

plot_race_lm <- function(data_list, race_acr, disease, study, to_plot = F) {
  
  city_sam_lr <- lm(data_list[[1]] ~ log(data_list[[2]]))
  race_total_pop_lr <- lm(data_list[[1]] ~ log(data_list[[3]]))
  Total_city_pop <- lm(data_list[[1]] ~ log(data_list[[4]]))
  
  if (to_plot) {
    plot(log(data_list[[2]]), data_list[[1]], main = paste(disease, ': Proportion Difference by Sample Size for ', race_acr, sep = ''), xlab = 'Log(City Sample Size)', ylab = 'Proportion Difference (Observed - Census Average)' )
    text(log(data_list[[2]]), data_list[[1]], 1:length(data_list[[2]]), pos=4, offset = 0.25)

  
    abline(city_sam_lr, col = 'red', lwd = 2)
  

    # mtext(paste(study, 'Adj R-squared=', round(summary(city_sam_lr)$adj.r.squared,3)), side = 3)
    mtext(paste(paste(study, collapse='+ ' ), 'Adj R-squared=', round(summary(city_sam_lr)$adj.r.squared,3)), side = 3)
    
    if(length(data_list[[1]]) > 2) {
      plot(city_sam_lr, 1)
      plot(city_sam_lr, 4)
    }
  
    plot(log(data_list[[3]]), data_list[[1]], main = paste(disease, ': Proportion Difference by City Population for ', race_acr, sep = ''), xlab = paste('log(',race_acr,' population)', sep = ''), ylab = 'Proportion Difference (Observed - Census Average)' )
    text(log(data_list[[3]]), data_list[[1]], 1:length(data_list[[3]]), pos=4, offset = 0.25)
  
  
    abline(race_total_pop_lr, col = 'red')
    mtext(paste(paste(study, collapse='+ ' ), 'Adj R-squared=', round(summary(race_total_pop_lr)$adj.r.squared,3)), side = 3)
    
    if (length(data_list[[1]]) > 2) {
      plot(race_total_pop_lr, 1)
      plot(race_total_pop_lr, 4)
    }
    
    
  
    plot(log(data_list[[4]]), data_list[[1]], main = paste(disease, ': Proportion Difference in ', race_acr, ' by Total City Population', sep = ''), xlab = 'log(Total City Population)', ylab = 'Proportion Difference (Observed - Census Average)' )
      text(log(data_list[[4]]), data_list[[1]], 1:length(data_list[[4]]), pos=4, offset = 0.25)
  
    
    abline(Total_city_pop, col = 'red')
    mtext(paste(paste(study, collapse='+ ' ), 'Adj R-squared=', round(summary(Total_city_pop)$adj.r.squared,3)), side = 3)
    
    if (length(data_list[[1]]) > 2) {
      plot(Total_city_pop, 1)
      plot(Total_city_pop, 4)
    }
  }

  
  
  return(list(city_sam_lr, race_total_pop_lr, Total_city_pop))
  
}
```

### Variables to loop through

```{r}
diseases <- c('SZSAFD', 'BAD12', 'CS')
col_pal <- c('#F8766D', '#7CAE00', '#00BFC4')
pch_codes <- c(7,9,10)
study1 <- 'GPC'
study2 <- c('COGS2', 'BSNIP1', 'BSNIP2')
```

## CA

### GPC
```{r}
coeffs_CA <- rep(0, length(diseases))
adjr2_CA <- rep(0, length(diseases))
p_vals_CA <- rep(0, length(diseases))

for (i in 1:length(diseases)) {
  rcp_d <- race_cities_plot_data2(df_avg, total_df, diseases[i],study1, 'CA')
  models <- plot_race_lm(rcp_d, 'CA', diseases[i], study1)
  coeffs_CA[i] <- summary(models[[3]])$coefficients[2]
  adjr2_CA[i] <- summary(models[[3]])$adj.r.squared
  p_vals_CA[i] <- summary(models[[3]])$coefficients[2,4]
}


names(coeffs_CA) <- diseases
coeffs_CA

adjr2_CA <- lapply(adjr2_CA, FUN = function(x) {x <- paste('adjr2 =', round(x,3))})

p_vals_CA <- lapply(p_vals_CA, FUN = function(x) { x <- paste('p =', round(x,4))})

bb_CA <- barplot(coeffs_CA, col = col_pal, 
                 ylab = 'Coefficient Estimate', 
                 ylim = c(min(coeffs_CA) - 0.01, 0.03),
                 main = 'GPC: Summary of Linear Regression Results for CA')
text(bb_CA,0 + 0.01,labels=adjr2_CA,cex=.8)
text(bb_CA, 0 + 0.003, labels=p_vals_CA, cex = 0.8)
```

### COGS2 + BSNIP
```{r}
coeffs_CA <- rep(0, length(diseases))
adjr2_CA <- rep(0, length(diseases))
p_vals_CA <- rep(0, length(diseases))

for (i in 1:length(diseases)) {
  rcp_d <- race_cities_plot_data2(df_avg, total_df, diseases[i],study2, 'CA')
  models <- plot_race_lm(rcp_d, 'CA', diseases[i], study2)
  coeffs_CA[i] <- summary(models[[3]])$coefficients[2]
  adjr2_CA[i] <- summary(models[[3]])$adj.r.squared
  p_vals_CA[i] <- summary(models[[3]])$coefficients[2,4]
}


names(coeffs_CA) <- diseases
coeffs_CA

adjr2_CA <- lapply(adjr2_CA, FUN = function(x) {x <- paste('adjr2 =', round(x,3))})

p_vals_CA <- lapply(p_vals_CA, FUN = function(x) { x <- paste('p =', round(x,4))})

bb_CA <- barplot(coeffs_CA, col = col_pal, 
                 ylab = 'Coefficient Estimate', 
                 ylim = c(min(coeffs_CA) - 0.01, 0.03),
                 main = 'COGS2 + BSNIP: Summary of Linear Regression Results for CA')
text(bb_CA,0 + 0.01,labels=adjr2_CA,cex=.8)
text(bb_CA, 0 + 0.003, labels=p_vals_CA, cex = 0.8)
```


## AS

### GPC
```{r}
coeffs_AS <- rep(0, length(diseases))
adjr2_AS <- rep(0, length(diseases))
p_vals_AS <- rep(0, length(diseases))

for (i in 1:length(diseases)) {
  rcp_d <- race_cities_plot_data2(df_avg, total_df, diseases[i],study1, 'AS')
  models <- plot_race_lm(rcp_d, 'AS', diseases[i], study1)
  coeffs_AS[i] <- summary(models[[3]])$coefficients[2]
  adjr2_AS[i] <- summary(models[[3]])$adj.r.squared
  p_vals_AS[i] <- summary(models[[3]])$coefficients[2,4]
}


names(coeffs_AS) <- diseases
coeffs_AS

adjr2_AS <- lapply(adjr2_AS, FUN = function(x) {x <- paste('adjr2 =', round(x,3))})

p_vals_AS <- lapply(p_vals_AS, FUN = function(x) { x <- paste('p =', round(x,4))})

bb_AS <- barplot(coeffs_AS, col = col_pal, 
                 ylab = 'Coefficient Estimate', 
                 ylim = c(min(coeffs_AS) - 0.01, 0.03),
                 main = 'GPC: Summary of Linear Regression Results for AS')
text(bb_AS,0 + 0.01,labels=adjr2_AS,cex=.8)
text(bb_AS, 0 + 0.003, labels=p_vals_AS, cex = 0.8)
```

### COGS2 + BSNIP
```{r}
coeffs_AS <- rep(0, length(diseases))
adjr2_AS <- rep(0, length(diseases))
p_vals_AS <- rep(0, length(diseases))

for (i in 1:length(diseases)) {
  rcp_d <- race_cities_plot_data2(df_avg, total_df, diseases[i],study2, 'AS')
  models <- plot_race_lm(rcp_d, 'AS', diseases[i], study2)
  coeffs_AS[i] <- summary(models[[3]])$coefficients[2]
  adjr2_AS[i] <- summary(models[[3]])$adj.r.squared
  p_vals_AS[i] <- summary(models[[3]])$coefficients[2,4]
}


names(coeffs_AS) <- diseases
coeffs_AS

adjr2_AS <- lapply(adjr2_AS, FUN = function(x) {x <- paste('adjr2 =', round(x,3))})

p_vals_AS <- lapply(p_vals_AS, FUN = function(x) { x <- paste('p =', round(x,4))})

bb_AS <- barplot(coeffs_AS, col = col_pal, 
                 ylab = 'Coefficient Estimate', 
                 ylim = c(min(coeffs_AS) - 0.01, 0.03),
                 main = 'COGS2 + BSNIP: Summary of Linear Regression Results for AS')
text(bb_AS,0 + 0.01,labels=adjr2_AS,cex=.8)
text(bb_AS, 0 + 0.003, labels=p_vals_AS, cex = 0.8)
```



## AA

### GPC
```{r}
coeffs_AA <- rep(0, length(diseases))
adjr2_AA <- rep(0, length(diseases))
p_vals_AA <- rep(0, length(diseases))

for (i in 1:length(diseases)) {
  rcp_d <- race_cities_plot_data2(df_avg, total_df, diseases[i],study1, 'AA')
  models <- plot_race_lm(rcp_d, 'AA', diseases[i], study1)
  coeffs_AA[i] <- summary(models[[3]])$coefficients[2]
  adjr2_AA[i] <- summary(models[[3]])$adj.r.squared
  p_vals_AA[i] <- summary(models[[3]])$coefficients[2,4]
}


names(coeffs_AA) <- diseases
coeffs_AA

adjr2_AA <- lapply(adjr2_AA, FUN = function(x) {x <- paste('adjr2 =', round(x,3))})

p_vals_AA <- lapply(p_vals_AA, FUN = function(x) { x <- paste('p =', round(x,4))})

bb_AA <- barplot(coeffs_AA, col = col_pal, 
                 ylab = 'Coefficient Estimate', 
                 ylim = c(min(coeffs_AA) - 0.01, 0.03),
                 main = 'GPC: Summary of Linear Regression Results for AA')
text(bb_AA,0 + 0.01,labels=adjr2_AA,cex=.8)
text(bb_AA, 0 + 0.003, labels=p_vals_AA, cex = 0.8)
```

### COGS2 + BSNIP
```{r}
coeffs_AA <- rep(0, length(diseases))
adjr2_AA <- rep(0, length(diseases))
p_vals_AA <- rep(0, length(diseases))

for (i in 1:length(diseases)) {
  rcp_d <- race_cities_plot_data2(df_avg, total_df, diseases[i],study2, 'AA')
  models <- plot_race_lm(rcp_d, 'AA', diseases[i], study2)
  coeffs_AA[i] <- summary(models[[3]])$coefficients[2]
  adjr2_AA[i] <- summary(models[[3]])$adj.r.squared
  p_vals_AA[i] <- summary(models[[3]])$coefficients[2,4]
}


names(coeffs_AA) <- diseases
coeffs_AA

adjr2_AA <- lapply(adjr2_AA, FUN = function(x) {x <- paste('adjr2 =', round(x,3))})

p_vals_AA <- lapply(p_vals_AA, FUN = function(x) { x <- paste('p =', round(x,4))})

bb_AA <- barplot(coeffs_AA, col = col_pal, 
                 ylab = 'Coefficient Estimate', 
                 ylim = c(min(coeffs_AA) - 0.01, 0.03),
                 main = 'COGS2 + BSNIP: Summary of Linear Regression Results for AA')
text(bb_AA,0 + 0.01,labels=adjr2_AA,cex=.8)
text(bb_AA, 0 + 0.003, labels=p_vals_AA, cex = 0.8)
```

## Alternate Solution:
- Response: Proportion difference
- Predictor1: log(Total City Population)
- Predictor2: indicator of in GPC or in COGS2 + BSNIP


## LRT pipeline

```{r}

LRT_pipeline <- function (dis, race_acr) {
  gpc_dat <- race_cities_plot_data2(df_avg, total_df, dis, study1, race_acr)
  y_gpc <- gpc_dat[['Prop dif']]
  x_gpc <- gpc_dat[['census average pop total']]
  
  cogs_bsnip_dat <- race_cities_plot_data2(df_avg, total_df, dis,study2, race_acr)
  
  y_cbs <- cogs_bsnip_dat[['Prop dif']]
  x_cbs <- cogs_bsnip_dat[['census average pop total']]
  
  df_for_lr <- data.frame(prop_dif = c(y_gpc, y_cbs),
                          avg_pop_total = c(x_gpc, x_cbs),
                          isGPC = c(rep(1, length(y_gpc)), rep(0,length(y_cbs))))
  
  full_mod<- lm(formula = prop_dif ~ log(avg_pop_total)*isGPC, data = df_for_lr)
  
  summary(full_mod)
  
  red_mod <- lm(prop_dif ~ log(avg_pop_total)*isGPC - log(avg_pop_total):isGPC, data = df_for_lr)
  
  summary(red_mod)
  
  # test to see if interaction term is important
  a1 <- anova(full_mod, red_mod, test = 'LRT')
  
  red_mod2 <- lm(prop_dif ~ log(avg_pop_total), data = df_for_lr)
  
  # test to see if GPC indicator is important
  a2 <- anova(red_mod, red_mod2, test = 'LRT')
  
  # test if full model fits significantly better than the fully reduced model
  a3 <- anova(full_mod, red_mod2, test = 'LRT')
  
  # test to see if added complexity needed
  a4 <- anova(red_mod2, test = 'LRT')
  
  p_vals <- c(a1$`Pr(>Chi)`, a2$`Pr(>Chi)`, a3$`Pr(>Chi)`, a4$`Pr(>F)`)
  p_vals <- p_vals[!is.na(p_vals)]
  
  res <- data.frame(test = c('Interaction Term', 'isGPC', 'Full Mod vs Reduced', 'Saturated vs Reduced'), p = p_vals)
  
  return(res)
                    
}
```

## LRT for AA
```{r}
for (i in 1:length(diseases)) {
  output <- list()
  output[[1]] <- paste('Condition: ', diseases[i], 'AA')
  output[[2]] <- LRT_pipeline(diseases[i], 'AA')
  print(output)
}
```

## LRT for AS
```{r}
for (i in 1:length(diseases)) {
  output <- list()
  output[[1]] <- paste('Condition: ', diseases[i], 'AS')
  output[[2]] <- LRT_pipeline(diseases[i], 'AS')
  print(output)
}
```

## LRT for CA
```{r}
for (i in 1:length(diseases)) {
  output <- list()
  output[[1]] <- paste('Condition: ', diseases[i], 'CA')
  output[[2]] <- LRT_pipeline(diseases[i], 'CA')
  print(output)
}
```
