==> Rmd Files/31. COGS2 and BSNIP AA Enrollment.Rmd <==
---
title: "COGS2 and BSNIP AA Enrollment"
author: "Daniel Zoleikhaeian"
date: "2023-05-15"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Loading libraries
```{r}
library(plyr)
library(dplyr)
library(ggplot2)
library(readxl)
```

# Loading dataset
```{r}
us_data <- read.csv('C:\\Users\\danie\\Documents\\Joshi Lab Materials\\3 Studies Dataset\\Dataset Merge\\bgc_merge_cDiag123_new.csv')
us_data$cEnrollmentYear <- as.numeric(substr(us_data$cEnrollmentDateYear, 1,4))

# Focusing only on AA
AA_sub <- us_data[us_data$cRace == 'AA', ]
```

# Finding the common cities between COGS2 and BSNIP
```{r}
bsnip_cities <- unique(AA_sub$cLocationCity[AA_sub$cStudy %in% c('BSNIP1', 'BSNIP2')])
gpc_cities <- unique(AA_sub$cLocationCity[AA_sub$cStudy== 'GPC'])
cogs2_cities <- unique(AA_sub$cLocationCity[AA_sub$cStudy == 'COGS2'])
com_cities <- intersect(bsnip_cities, cogs2_cities)
com_cities # no cities in common between COGS2 and BSNIP

intersect(bsnip_cities, gpc_cities) # no cities in common between BSNIP and GPC

cogs2gpc_com <- intersect(cogs2_cities, gpc_cities) # two cities in common between COGS2 and GPC

# Creates problem for "within-city-variance"

bsnip1_cities <- unique(AA_sub$cLocationCity[AA_sub$cStudy == 'BSNIP1'])
bsnip2_cities <- unique(AA_sub$cLocationCity[AA_sub$cStudy == 'BSNIP2'])
bsnip12_com <- intersect(bsnip1_cities, bsnip2_cities)
bsnip12_com # 4 cities in common between bsnip1 and bsnip2

```

* Conclusions:
  * Lack of intersection among cities between studies makes it difficult to get a "within-city-variance"
  * No cities in common between GPC and BSNIP12
  * No cities in common between COGS2 and BSNIP12


# Helper Function: Counts to proportions within a city within a year
```{r}
# vectorized helper function to deal with NA
is_true <- Vectorize(isTRUE)

count2prop <- function(df_in, year) {
  race_ct <- plyr::count(df_in[is_true(df_in$cEnrollmentYear == year), ], 'cRace')
  race_ct$prop <- race_ct$freq / sum(race_ct$freq)
  return(race_ct)
}
```

## FINDING: incomplete cases create issues for dataframe subsetting
* Lots of missing enrollment dates
```{r}
inc_cases <- us_data[!complete.cases(us_data),]

# total number of incomplete cases
nrow(inc_cases)

# total number of incomplete cases due to missing enrollment date
sum(is.na(inc_cases$cEnrollmentYear))

sum(inc_cases$cLocationCity == 'Los Angeles' & inc_cases$cStudy == 'GPC')
nrow(us_data[is_true(us_data$cLocationCity == 'Los Angeles') & is_true(us_data$cStudy == 'GPC'),])

# checking the cases that aren't due to missing enrollment year
View(inc_cases[!is.na(inc_cases$cEnrollmentYear), ])
## NOTE: none from COGS2

# just missing age
nrow(us_data)

# seeing how many missing cases in each study
inc_counts <- plyr::count(inc_cases, 'cStudy')

# comparing to total cases
total_counts <- plyr::count(us_data, 'cStudy')
## BSNIP1 had no missing data

# creating a proportion
inc_counts$proportion <- inc_counts$freq/total_counts$freq[2:4]
inc_counts

## FINDING 11% missing in COGS2

write.csv(inc_cases, file = "C:/Users/danie/Documents/Joshi Lab Materials/3 Studies Dataset/Dataset Merge/incomplete_cases_us.csv",row.names = F)
```

# Making histograms: one for each city
* Color by study
* x-axis is year
* y-axis is AA count

## For cities in BSNIP1 and BSNIP2 (using raw counts)
```{r}
for (i in 1:length(bsnip12_com)) {
  study_sub <- AA_sub[AA_sub$cLocationCity == bsnip12_com[i] & AA_sub$cStudy == 'BSNIP1', ]
  hist(as.numeric(study_sub$cEnrollmentYear), breaks = length(unique(study_sub$cEnrollmentYear)), xlim = c(2008, 2020), ylim=c(0,80),
       main = paste('AA Enrollment in', bsnip12_com[i], 'for BSNIP1 BSNIP2'),
       xlab = 'Year',
       ylab = 'Count')
  
  study_sub <- AA_sub[AA_sub$cLocationCity == bsnip12_com[i] & AA_sub$cStudy == 'BSNIP2', ]
  hist(as.numeric(study_sub$cEnrollmentYear), breaks = length(unique(study_sub$cEnrollmentYear)), xlim = c(2008, 2020), add = TRUE, col= 'red')
  
  legend('topleft', fill = c('gray', 'red'), legend=c('BSNIP1', 'BSNIP2'))
}
```

### using proportions instead
```{r}
years <- unique(us_data$cEnrollmentYear[us_data$cStudy %in% c('BSNIP1', 'BSNIP2')]) 
AA_prop <- rep(0, length(years))
AA_prop2 <- rep(0, length(years))
bsnip_dat <- us_data[us_data$cStudy %in% c('BSNIP1', 'BSNIP2'),]

for(i in 1:length(bsnip12_com)) {
  for(j in 1:length(AA_prop)) {
    AA_prop[j] <- count2prop(bsnip_dat[is_true(bsnip_dat$cLocationCity == bsnip12_com[i]),] ,years[j])$prop[1]
    AA_prop2[j]
  }
  plot(years, AA_prop, 
         main = paste('AA Enrollment in', bsnip12_com[i], 'for BSNIP1 BSNIP2'),
         xlab = "Year",
         ylab = 'AA Proportion',
       ylim = c(0, 1),
       xlim = c(2008,2020), pch = 19, col = 'blue'
       )
  grid()
}
```

## For cities in GPC and COGS2 (using raw counts)
```{r}
for (i in 1:length(cogs2gpc_com)) {
  study_sub <- AA_sub[AA_sub$cLocationCity == cogs2gpc_com[i] & AA_sub$cStudy == 'GPC', ]
  hist(as.numeric(study_sub$cEnrollmentYear), breaks = length(unique(study_sub$cEnrollmentYear)), xlim = c(2008, 2020),
       main = paste('AA Enrollment in', cogs2gpc_com[i], 'for GPC COGS2'),
       xlab = 'Year',
       ylab = 'Count', col = 'darkgreen')
  
  study_sub <- AA_sub[AA_sub$cLocationCity == cogs2gpc_com[i] & AA_sub$cStudy == 'COGS2', ]
  hist(as.numeric(study_sub$cEnrollmentYear), breaks = length(unique(study_sub$cEnrollmentYear)), xlim = c(2008, 2020), add = TRUE, col= 'orange')
  
  legend('topleft', fill = c('darkgreen', 'orange'), legend=c('GPC', 'COGS2'))
}
```
* Uninformative due to large sample size difference

### using proportions instead
```{r}
years <- unique(AA_sub$cEnrollmentYear[AA_sub$cStudy %in% c('COGS2', 'GPC')])
years <- years[!is.na(years)]

AA_prop <- rep(0, length(years))
AA_prop2 <- rep(0, length(years))

cogs2_dat <- us_data[is_true(us_data$cStudy == 'COGS2'),]
gpc_dat <- us_data[is_true(us_data$cStudy == 'GPC'), ]

# for every city in common between gpc and cogs2
for(i in 1:length(cogs2gpc_com)) {

# get the AA proportion for that year in that icty
  for(j in 1:length(AA_prop)) {
    
    # subset the city
    gpc_sub <- gpc_dat[is_true(gpc_dat$cLocationCity == cogs2gpc_com[i]),]
    
    # get the race proportions
    gpc_c2p <- count2prop(gpc_sub ,years[j])
    gpc_c2p <- gpc_c2p[!is.na(gpc_c2p$cRace),]
    
    # extract the proportion corresponding to AA
    gpc_v_i <- gpc_c2p$prop[gpc_c2p$cRace == 'AA']
    
    cogs2_sub <- cogs2_dat[is_true(cogs2_dat$cLocationCity == cogs2gpc_com[i]),]
    cogs2_c2p <- count2prop(cogs2_sub ,years[j])
    cogs2_c2p <- cogs2_c2p[!is.na(cogs2_c2p$cRace),]
    cogs2_v_i <- cogs2_c2p$prop[cogs2_c2p$cRace == 'AA']
    

    if (length(gpc_v_i) != 0) {
       AA_prop[j] <- gpc_v_i
    }
    if (length(cogs2_v_i) != 0) {
        AA_prop2[j] <- cogs2_v_i
    }
  }
  plot(years, AA_prop, 
         main = paste('AA Enrollment in', cogs2gpc_com[i], 'for GPC COGS2'),
         xlab = "Year",
         ylab = 'AA Proportion',
       ylim = c(0, 1),
       xlim = c(2008,2020), col = 'darkgreen', pch =19)
  points(x = years, y = AA_prop2, col = 'orange', pch =19)
  legend('topleft', legend=c('GPC', 'COGS2'), col = c('darkgreen','orange'), pch = 19)
  grid()
  print(AA_prop)
  # print(AA_prop2)
}
```
