==> Rmd Files/37. Gender COGS2.Rmd <==
---
title: "37. Gender HC + SZSAFD"
author: "Daniel Zoleikhaeian"
date: "2023-06-12"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Importing Dataset
```{r}
library(plyr)
library(dplyr)
library(ggplot2)
df_COGS2 <- read.csv('C:\\Users\\danie\\Documents\\Joshi Lab Materials\\3 Studies Dataset\\Dataset Merge\\bgc_merge_cDiag123_060723.csv')
df_COGS2 <- df_COGS2[df_COGS2$cStudy == 'COGS2', ]

# adding year column
df_COGS2$cEnrollmentYear <- as.numeric(substr(df_COGS2$cEnrollmentDateYear, 1,4))

# no NAs; dataset is complete
nrow(df_COGS2) == sum(complete.cases(df_COGS2)) # TRUE
```

# Helper functions

## By-city proportion generator
```{r}
prop_grt <- function(df) {
  props <- rep(0, nrow(df))
  for (i in 1:nrow(df)) {
    props[i] <- df$freq[i] / sum(df$freq[df$cLocationCity == df$cLocationCity[i]])
  }
  df$props <- props
  return (df)
}
```

### Gender plotter
```{r}
gen_plottr <- function(df, grp_name) {
  ggplot(data=df, aes(x = cLocationCity, y = props, fill = cGender)) + geom_bar(stat = 'identity') + ylab('Proportion') + ggtitle(paste('COGS2: ', grp_name, ' Gender Proportions by City', sep = ''))
}
```

### Gender plotter - boxplot
```{r}
gen_box <- function(df, grp_name) {
  df_sub <- df[df$cGender == grp_name, ]
  ggplot(data = df_sub, x = Diagnosis, y = props) + geom_boxplot(aes(x = Diagnosis, y = props)) + geom_point(aes(x = Diagnosis, y = props, color = cLocationCity), size = 3, shape = 19) + ggtitle(paste('COGS2 ', grp_name, ' Proportions across Cities', sep = '')) +  ylab('Proportion')
}
```

# Analyzing Gender by City

## Time-invariant exploration

### All subjects
```{r}
gen_counts <- plyr::count(df_COGS2, vars = c('cLocationCity', 'cGender'))
#View(gen_counts)

gen_counts <- prop_grt(gen_counts)

gp_all <- gen_plottr(gen_counts, 'Total')
gp_all
```

### HC
```{r}
hc_gen <- plyr::count(df_COGS2[df_COGS2$cDiagnosis4 == 'CS',], vars = c('cLocationCity', 'cGender'))
#View(hc_gen)

hc_gen <- prop_grt(hc_gen)

gp_hc <- gen_plottr(hc_gen, 'CS')
gp_hc
```

### SZSAFD
```{r}
szsafd_gen <- plyr::count(df_COGS2[df_COGS2$cDiagnosis4 == 'SZSAFD',], vars = c('cLocationCity', 'cGender'))
#View(szsafd_gen)

szsafd_gen <- prop_grt(szsafd_gen)

gp_szsafd <- gen_plottr(szsafd_gen, 'SZSAFD')
gp_szsafd
```

### Boxplot representations
```{r}
hc_gen$Diagnosis <- rep('CS', nrow(hc_gen))
szsafd_gen$Diagnosis <- rep('SZSAFD', nrow(szsafd_gen))
total_df <- rbind(hc_gen, szsafd_gen)
box_gp_m <- gen_box(total_df, 'M')
box_gp_f <- gen_box(total_df, 'F')
box_gp_m
box_gp_f
```

## Time-variant exploration - Gender proportions over time (cumulative)

### Generating proportions
```{r}
genders <- c('M', 'F')
diagnoses <- c('CS', 'SZSAFD')
cities <- unique(df_COGS2$cLocationCity)
years <- unique(df_COGS2$cEnrollmentYear)

n <- 2 * length(diagnoses) * length(cities) * length(years)

# empty dataframe to hold data
df_gen <- data.frame(City = rep('',n),
                    Year = rep(0,n),
                    Diagnosis = rep('',n),
                    Gender = rep('', n),
                    Prop = rep(0, n)
                    )
n_track <- 1
for (d in 1:length(diagnoses)) {
  for (i in 1:length(years)) {
    for (j in 1:length(cities)) {
      df_sub <- df_COGS2[df_COGS2$cDiagnosis3 == diagnoses[d] & df_COGS2$cLocationCity == cities[j] & df_COGS2$cEnrollmentYear <= years[i],]
      if (nrow(df_sub) == 0) {
        n_track <- n_track + 2
      next
    }
      gen_cts <- plyr::count(df_sub, var = 'cGender')
      gen_cts$props <- gen_cts$freq / sum(gen_cts$freq)
      
      df_gen[n_track:(n_track+1),1] <- cities[j]
      df_gen[n_track:(n_track+1),2] <- years[i]
      df_gen[n_track:(n_track+1),3] <- diagnoses[d]
      df_gen[n_track:(n_track+1),4] <- gen_cts$cGender
      df_gen[n_track:(n_track+1), 5] <- gen_cts$props 
      n_track <- n_track + 2
    }
  }
}
df_gen
```

### Plotting the proportions
```{r}
#View(males)
males <- df_gen[df_gen$Gender == 'M', ]
females <- df_gen[df_gen$Gender == 'F', ]
# nrow(males) + nrow(females) == nrow(df_gen) # TRUE

gm_p <- ggplot(data=males, aes(x=Year, y=Prop, group=interaction(City, Diagnosis), linetype = Diagnosis)) +
  geom_line(data = males, aes(color = City), size = 1.1) +
  geom_point(aes(color = City)) +
  ggtitle('COGS2: Male Cumulative Proportions over\nTime for CS and SZSAFD') + ylab('Proportion') + theme(plot.title = element_text(hjust = 0.5))

gf_p <- ggplot(data=females, aes(x=Year, y=Prop, group=interaction(City, Diagnosis), linetype = Diagnosis)) +
  geom_line(data = females, aes(color = City), size = 1.1) +
  geom_point(aes(color = City)) +
  ggtitle('COGS2: Cumulative Female Proportions over\nTime for CS and SZSAFD') + ylab('Proportion') + theme(plot.title = element_text(hjust = 0.5))

gm_p
gf_p
```

Preliminary conclusions:
* Higher proportion of SZSAFD males across all cities
* No clear convergence towards equal split in either CS or SZSAFD
