==> Rmd Files/39. Diversity Scores COGS2 - Race, Gender, Ethnicity.Rmd <==
---
title: "39. Diversity Index COGS2 - Race, Gender, Ethnicity"
author: "Daniel Zoleikhaeian"
date: "2023-06-12"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Importing data
```{r}
library(plyr)
library(dplyr)
library(ggplot2)
df_COGS2 <- read.csv('C:\\Users\\danie\\Documents\\Joshi Lab Materials\\3 Studies Dataset\\Dataset Merge\\bgc_merge_cDiag123_060723.csv')
df_COGS2 <- df_COGS2[df_COGS2$cStudy == 'COGS2', ]
nrow(df_COGS2)

head(df_COGS2)
# adding year column
df_COGS2$cEnrollmentYear <- as.numeric(substr(df_COGS2$cEnrollmentDateYear, 1,4))
head(df_COGS2)

unique(df_COGS2$cRace)

# re-encoding OT/UNK/MR into one group
df_COGS2$cRace2 <- df_COGS2$cRace
df_COGS2$cRace2[df_COGS2$cRace2 %in% c('OT', 'OT/UNK', 'MR')] <- 'OT/MR'
nrow(df_COGS2)
```

### Helper function: Diversity Index
```{r}
mult_ent <- function(race_prop_vec) {
  
  tot <- 0
  
  for (i in 1:length(race_prop_vec)) {
    if (race_prop_vec[i] != 0) {
      tot <- tot + race_prop_vec[i] * log(1/race_prop_vec[i])
    }
  }
  return(tot)
}
```

## Checking sparseness of categories
```{r}

df_count_og <- plyr::count(df_COGS2, vars=c('cRace2', 'cHispanicorLatino', 'cGender'))
df_count_og
length(unique(df_COGS2$cRace2)) # 7 race categories
# expected: 7 races * 2 ethnicities * 2 gender = 28 categories
nrow(df_count_og) # only 25 categories here

## Suggestion: remove the binary category that is too sparse
df_count_2 <- plyr::count(df_COGS2, vars=c('cRace2', 'cGender'))
nrow(df_count_2) # 14 rows, as expected

df_count_3 <- plyr::count(df_COGS2, vars=c('cRace2', 'cHispanicorLatino'))
nrow(df_count_3) # 13 rows; this was the sparse category

# max diversity score will use 28 categories
```

## Generating Diversity Scores (Cumulative) by Year and City
```{r}
hist(df_COGS2$cAge[df_COGS2$cDiagnosis3 == 'SZSAFD'])
hist(df_COGS2$cAge[df_COGS2$cDiagnosis3 == 'CS'], add=T, col = rgb(1, 0, 0, alpha = 0.5))

diagnoses <- c('CS', 'SZSAFD')
cities <- unique(df_COGS2$cLocationCity)
#years <- unique(df_COGS2$cEnrollmentYear)
years <- 2011:2014

n <- 2 * length(cities) * length(years)

# empty dataframe to hold data
df_di <- data.frame(City = rep('',n),
                    Year = rep(0,n),
                    Diagnosis = rep('',n),
                    mult_ent = rep(0,n)
                    )

# tertiary level: diagnosis
# outer level: year
# inner level: city
n_track <- 1
for (d in 1:length(diagnoses)) {
  for (i in 1:length(years)) {
    for (j in 1:length(cities)) {
      df_sub <- df_COGS2[df_COGS2$cDiagnosis3 == diagnoses[d] & df_COGS2$cLocationCity == cities[j] & df_COGS2$cEnrollmentYear <= years[i],]
      
      # if the dataframe is empty, skip this iteration
      if (nrow(df_sub) == 0) {
        df_di[n_track,1] <- cities[j]
        df_di[n_track,2] <- years[i]
        df_di[n_track,3] <- diagnoses[d]
        df_di[n_track,4] <- -99 # code for no data
        n_track <- n_track + 1
        next
      }
      
      group_ct <- plyr::count(df_sub, c('cRace2', 'cGender', 'cHispanicorLatino'))
      group_ct$prop <- group_ct$freq/nrow(df_sub)
      
      prop_vec <- group_ct$prop
      m_ent <- mult_ent(prop_vec)
      
      df_di[n_track,1] <- cities[j]
      df_di[n_track,2] <- years[i]
      df_di[n_track,3] <- diagnoses[d]
      df_di[n_track,4] <- m_ent
      n_track <- n_track+1
    }
  }
}

View(df_di)

max_di <- log(28) # 3.332205

# rescaling multigroup entropy

df_di$mult_ent_s <- df_di$mult_ent / max_di * 100
```

## Plotting Diversity Scores

### Time-invariant analysis
```{r}
df_2014 <- df_di[df_di$Year == 2014, ]
View(df_2014)
bar_DI <- ggplot(data = df_2014, aes(x = City, y = mult_ent, fill = Diagnosis)) + geom_bar(stat = 'identity', position = position_dodge()) + ylab('Diversity Index') + ggtitle('Diversity Index for CS and SZSAFD\n across Cities')
bar_DI

box_DI <- ggplot(data = df_2014, x = Diagnosis, y = mult_ent) + geom_boxplot(aes(x = Diagnosis, y = mult_ent)) + geom_point(aes(x = Diagnosis, y = mult_ent, color = City), size = 3, shape = 19) + ggtitle('Diversity Indexes across Cities') +  ylab('Diversity Index')
box_DI

box_DI_s <- ggplot(data = df_2014, x = Diagnosis, y = mult_ent_s) + geom_boxplot(aes(x = Diagnosis, y = mult_ent_s)) + geom_point(aes(x = Diagnosis, y = mult_ent_s, color = City), size = 3, shape = 19) + ggtitle('Diversity Indexes across Cities') +  ylab('Percent of Max Diversity Index')

box_DI_s
```

Initial thoughts for testing difference:
Weighted two-sample t-test
* Issues: 
- sample sizes are too small to assume normal distribution (n = 5)
- groups (CS and SZSAFD) are not independent
Suggestion: just note ranges 
* Difficulty: the cities are not consistent
- No clear relationship between CS diversity index and SZSAFD diversity index

### Time-variant analysis
```{r}
library(ggplot2)

df_di_cs <- df_di[df_di$Diagnosis == 'CS',]
df_di_szsafd <- df_di[df_di$Diagnosis == 'SZSAFD',]

ggplot(data=df_di, aes(x=Year, y=mult_ent, group=interaction(City, Diagnosis), linetype = Diagnosis)) +
  geom_line(data = df_di, aes(color = City), size = 1.1) +
  geom_point(aes(color = City)) +
  ggtitle('Cumulative Diversity Indices\nover Time for HC (solid) and SZSAFD (dashed)') + ylab('Diversity Index')

# same thing but percentage of max_di
ggplot(data=df_di, aes(x=Year, y=mult_ent_s, group=interaction(City, Diagnosis), linetype = Diagnosis)) +
  geom_line(data = df_di, aes(color = City), size = 1.1) +
  geom_point(aes(color = City)) +
  ggtitle('Cumulative Diversity Indices\nover Time for HC (solid) and SZSAFD (dashed)') + ylab('Percent of Maximum Diversity Index')
```

## Longitudinal Data Analysis - Framing the question
* Goal: predict diversity index in sample based on:
1) City diversity index - covariate
2) City population (estimate from ACS) - covariate
3) Cumulative Sample Size from that city - covariate
4) Diagnosis designation (variable of interest)

Resource: 

Lalonde, T.L. (2015). Modeling Time-Dependent Covariates in Longitudinal Data Analyses. In: Chen, DG., Wilson, J. (eds) Innovative Statistical Methods for Public Health Data. ICSA Book Series in Statistics. Springer, Cham. https://doi.org/10.1007/978-3-319-18536-1_4

From the graph:
* Can test a diagnosis x time effect within each city 
- See if the responses are parallel or not
* Can see how diversity index profiles vary based on city characteristics

## Testing Diagnosis x time effect
NOTE: no time-dependent covariates involved in model
```{r}
library(nlme)
library(lme4)

# transformation: years post enrollment date
df_di$Post10 <- df_di$Year-2010

# random intercept and slopes model does not converge

# use random intercept model instead
ri_mod <- lme(fixed = mult_ent_s ~ Diagnosis * factor(Post10), random = ~ 1|City,
                  method = "REML", na.action = na.omit, data=df_di)
summary(ri_mod)

ri_mod_ct <- lme(fixed = mult_ent_s ~ Diagnosis * Post10, random = ~ 1|City,
                  method = "REML", na.action = na.omit, data=df_di)
summary(ri_mod_ct)

ri_mod_ml <- lme(fixed = mult_ent_s ~ Diagnosis * factor(Post10), random = ~ 1|City,
                  method = "ML", na.action = na.omit, data=df_di)

ri_mod_ct_ml <- lme(fixed = mult_ent_s ~ Diagnosis * Post10, random = ~ 1|City,
                  method = "ML", na.action = na.omit, data=df_di)

anova(ri_mod_ct_ml, ri_mod_ml)

# no evidence for response profile fitting significantly better than continuous time
summary(ri_mod_ct)
```

Preliminary conclusion:

There is a significant increase in diversity score over time. On average, the diversity score became 2.78433 percentage points closer to the maximum diversity score (p = 0.0002). At baseline, the diversity scores between SZSAFD and CS groups were non-significant (p = 0.9587). There was a slight attenuation in the increase in diversity scores over time for the SZSAFD group (-0.83114), but this attenuation was non-significant (p = 0.3849).
