==> Rmd Files/41. City vs Study DI-revised agg.Rmd <==
---
title: "Census vs Study Diversity Indices"
author: "Daniel Zoleikhaeian"
date: "2023-06-21"
output:
  pdf_document: default
  html_document: default
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, tidy.opts=list(width.cutoff=60), tidy=TRUE)
```

# Setup

## Importing data
```{r}
library(grid)
library(plyr)
library(dplyr)
library(ggplot2)
library(gridExtra)
df_COGS2 <- read.csv('C:\\Users\\danie\\Documents\\Joshi Lab Materials\\3 Studies Dataset\\Dataset Merge\\bgc_merge_cDiag123_060723.csv')
df_COGS2 <- df_COGS2[df_COGS2$cStudy == 'COGS2', ]
nrow(df_COGS2)

head(df_COGS2)
# adding year column
df_COGS2$cEnrollmentYear <- as.numeric(substr(df_COGS2$cEnrollmentDateYear, 1,4))
head(df_COGS2)

unique(df_COGS2$cRace)

# re-encoding OT/UNK/MR into one group
df_COGS2$cRace2 <- df_COGS2$cRace
df_COGS2$cRace2[df_COGS2$cRace2 %in% c('OT', 'OT/UNK', 'MR', 'UNK')] <- 'OT/MR'
nrow(df_COGS2)
```

### Helper function: Diversity Index
```{r}
mult_ent <- function(race_prop_vec) {
  
  tot <- 0
  
  for (i in 1:length(race_prop_vec)) {
    if (race_prop_vec[i] != 0) {
      tot <- tot + race_prop_vec[i] * log(1/race_prop_vec[i])
    }
  }
  return(tot)
}
```


# COGS2 Analysis + Plots

## COGS2: Aggregate and By-City DI

### Aggregate DI
```{r}
#pdf(file='Census_vs_Study_div_index.pdf')
head(df_COGS2)
group_ct <- plyr::count(df_COGS2, c('cRace2', 'cGender', 'cHispanicorLatino'))
group_ct$prop <- group_ct$freq/nrow(df_COGS2)
prop_vec <- group_ct$prop
prop_vec
agg_m_ent <- mult_ent(prop_vec)
agg_m_ent

agg_df <- data.frame(City = 'COGS2 Aggregate', mult_ent = agg_m_ent)

# same thing but without SD
no_sd <- df_COGS2[df_COGS2$cLocationCity != 'San Diego', ]
group_ct <- plyr::count(no_sd, c('cRace2', 'cGender', 'cHispanicorLatino'))
group_ct$prop <- group_ct$freq/nrow(no_sd)
prop_vec <- group_ct$prop
prop_vec
agg_m_ent_nosd <- mult_ent(prop_vec)
agg_m_ent_nosd
agg_df_nosd <- data.frame(City = 'COGS2 No SD', mult_ent = agg_m_ent_nosd)

```

### By-City DI
```{r}
cities <- unique(df_COGS2$cLocationCity)
n <- length(cities)
df_di <- data.frame(City = rep('',n),
                    mult_ent = rep(0,n)
                    )
head(df_di)
for (i in 1:length(cities)) {
  df_sub <- df_COGS2[df_COGS2$cLocationCity == cities[i],]
  
  group_ct <- plyr::count(df_sub, c('cRace2', 'cGender', 'cHispanicorLatino'))
  group_ct$prop <- group_ct$freq/nrow(df_sub)
      
  prop_vec <- group_ct$prop
  m_ent <- mult_ent(prop_vec)
  
  df_di[i,1] <- cities[i]
  df_di[i,2] <- m_ent
}
head(df_di)


df_di <- rbind(df_di,agg_df, agg_df_nosd)
df_di
```

### Barplot comparing results
```{r}
my_cols <- c('black', 'darkgray', 'chartreuse3', 'brown3', 'cornflowerblue', 'darkgoldenrod2', 'darkorchid4')

cogs2_bar <- ggplot(data = df_di, aes(x = City, y = mult_ent, fill = City)) + geom_bar(stat = 'identity', position = position_dodge()) + ylab('Diversity Index') + ggtitle('Diversity Index COGS2 Cities') + theme(axis.text.x = element_blank()) + 
  scale_fill_manual(values=my_cols)
cogs2_bar
```

## COGS2: CS and SZSAFD split
```{r}
diagnoses <- c('CS', 'SZSAFD')
cities2 <- c(cities, 'Aggregate')
n <- 2 * length(cities2)

df_di2 <- data.frame(City = rep('',n),
                    Diagnosis = rep('',n),
                    mult_ent = rep(0,n)
                    )
n_track <- 1
for (d in 1:length(diagnoses)) {
    for (j in 1:length(cities2)) {
      if (cities2[j] != 'Aggregate') {
        df_sub <- df_COGS2[df_COGS2$cDiagnosis3 == diagnoses[d] & df_COGS2$cLocationCity == cities2[j],]
      } else {
        df_sub <- df_COGS2[df_COGS2$cDiagnosis3 == diagnoses[d],]
      }
      
      
      # if the dataframe is empty, skip this iteration
      if (nrow(df_sub) == 0) {
        df_di2[n_track,1] <- cities2[j]
        df_di2[n_track,2] <- diagnoses[d]
        df_di2[n_track,3] <- -99 # code for no data
        n_track <- n_track + 1
        next
      }
      
      group_ct <- plyr::count(df_sub, c('cRace2', 'cGender', 'cHispanicorLatino'))
      group_ct$prop <- group_ct$freq/nrow(df_sub)
      
      prop_vec <- group_ct$prop
      m_ent <- mult_ent(prop_vec)
      
      df_di2[n_track,1] <- cities2[j]
      df_di2[n_track,2] <- diagnoses[d]
      df_di2[n_track,3] <- m_ent
      n_track <- n_track+1
    }
}

#View(df_di2)
```

### Barplot
```{r}
bar2_COGS2 <- ggplot(data = df_di2, aes(x = City, y = mult_ent, fill = Diagnosis)) + geom_bar(stat = 'identity', position = position_dodge()) + ylab('Diversity Index') + ggtitle('Diversity Index COGS2 Cities') 
bar2_COGS2
```

### Same thing but without San Diego
```{r}
cities3 <- cities2[-1]

n <- 2 * length(cities3)

df_di3 <- data.frame(City = rep('',n),
                    Diagnosis = rep('',n),
                    mult_ent = rep(0,n)
                    )
n_track <- 1
for (d in 1:length(diagnoses)) {
    for (j in 1:length(cities3)) {
      if (cities3[j] != 'Aggregate') {
        df_sub <- no_sd[no_sd$cDiagnosis3 == diagnoses[d] & no_sd$cLocationCity == cities3[j],]
      } else {
        df_sub <- no_sd[no_sd$cDiagnosis3 == diagnoses[d],]
      }
      
      
      # if the dataframe is empty, skip this iteration
      if (nrow(df_sub) == 0) {
        df_di3[n_track,1] <- cities3[j]
        df_di3[n_track,2] <- diagnoses[d]
        df_di3[n_track,3] <- -99 # code for no data
        n_track <- n_track + 1
        next
      }
      
      group_ct <- plyr::count(df_sub, c('cRace2', 'cGender', 'cHispanicorLatino'))
      group_ct$prop <- group_ct$freq/nrow(df_sub)
      
      prop_vec <- group_ct$prop
      m_ent <- mult_ent(prop_vec)
      
      df_di3[n_track,1] <- cities3[j]
      df_di3[n_track,2] <- diagnoses[d]
      df_di3[n_track,3] <- m_ent
      n_track <- n_track+1
    }
}

bar3_COGS2 <- ggplot(data = df_di3, aes(x = City, y = mult_ent, fill = Diagnosis)) + geom_bar(stat = 'identity', position = position_dodge()) + ylab('Diversity Index') + ggtitle('Diversity Index COGS2 Cities - No San Diego') 
bar3_COGS2

plyr::count(df_COGS2, 'cLocationCity') # losing 576 samples by ignoring SD
```

# ACS Analysis

## Importing and checking data
```{r}
df_acs <- read.csv('C:\\Users\\danie\\Documents\\Joshi Lab Materials\\acs_cogs2_1014.csv')
unique(df_acs$CITY) # all cities except SD
head(df_acs)

# no missing data
sum(complete.cases(df_acs)) == nrow(df_acs)

```

## Encoding new race categories, binarizing hispan category
```{r}
df_acs$Race2 <- rep(0, nrow(df_acs))
df_acs$Hispan2 <- rep(0, nrow(df_acs))

sum(df_acs$HISPAN == 9) # everyone reported a hispanic status
df_acs$Hispan2 <- as.numeric(df_acs$HISPAN != 0) # 0 for not hispanic or latino, else 1

PI_raced <- c(680:699) # PI races

df_acs$Race2[df_acs$RACE == 1 ] <- 1 # White
df_acs$Race2[df_acs$RACE == 2 ] <- 2 # Black
df_acs$Race2[df_acs$RACE == 3 ] <- 3 # American Indian or Alaska Native
df_acs$Race2[df_acs$RACE %in% 4:6 & !(df_acs$RACED %in% PI_raced) ] <- 4 # Asian
df_acs$Race2[df_acs$RACE == 6 & df_acs$RACED %in% PI_raced ] <- 5 # Pacific Islander (or Native Hawaiian)
df_acs$Race2[df_acs$RACE %in% 7:9 ] <- 6 # Mixed/Other
```

## Comparing ACS age ranges to COGS2 age ranges
```{r}
# Getting the counts from no_sd
city_sam_sizes <- plyr::count(no_sd, 'cLocationCity')
city_sam_sizes

# age range of COGS2 was 18-65
hist(no_sd$cAge)
summary(no_sd$cAge)

# age range of the ipums sample: 0-95
hist(df_acs$AGE)
summary(df_acs$AGE)

# solution: truncate the acs dataframe by the age in COGS2
acs1865 <- df_acs[df_acs$AGE >= 18 & df_acs$AGE <= 65, ]
 ( nrow(df_acs) - nrow(acs1865) ) / nrow(df_acs) * 100 # lost 34% of the rows
```

## Calculating diversity index for ACS sample
```{r}

# effective total population
# note: due to truncation, city pop is no longer relevant

# empty dataframe to hold results
acs_di <- data.frame(CITY = rep('', 4),
                     DI = rep(0, 4))

cogs_cities_acs <- unique(acs1865$CITY)
cogs_cities_acs2 <- c(-1,cogs_cities_acs) # -1 is surrogate for aggregate

for (i in 1:length(cogs_cities_acs2)) {
  if (cogs_cities_acs2[i] != -1) {
    df_sub <- acs1865[acs1865$CITY == cogs_cities_acs2[i], ]
  } else {
    df_sub <- acs1865
  }
  
  # total effective population for that city
  # need to use this cuz subset by age
  tot <- sum(df_sub$PERWT)
  
  weighted_cts <- plyr::count(df_sub, c('Race2', 'Hispan2', 'SEX'), wt_var = 'PERWT')
  
  props <- weighted_cts$freq / tot
  acs_di[i,1] <-  cogs_cities_acs2[i]
  acs_di[i,2] <- mult_ent(props)
}

#View(acs_di)

fac_test <- factor(acs_di$CITY)
levels(fac_test) <- c('Aggregate', 'Los Angeles', 'New York', 'Philadelphia', 'Seattle')
fac_test


acs_di$City <- fac_test
acs_di <- acs_di[, c('City', 'CITY', 'DI')]
colnames(acs_di)[2] <- 'City_code'

knitr::kable(acs_di)
grid.newpage()
grid.table(acs_di, rows = NULL)

# showing populations
pop_counts <- plyr::count(acs1865, 'CITY', 'PERWT')
pop_counts$City <- c('Los Angeles', 'New York', 'Philadelphia', 'Seattle')
pop_counts[, c('City', 'CITY', 'freq')]

# LA and NY have highest counts
```

# Analysis: Comparing ACS DI to study DI
```{r}
# putting the dataframes on top of each other
cs_cogs2 <- df_di3[df_di3$Diagnosis == 'CS',]
szsafd_cogs2 <- df_di3[df_di3$Diagnosis == 'SZSAFD',]
acs_di$Diagnosis <- 'Census ACS'
colnames(cs_cogs2)[3] = colnames(szsafd_cogs2)[3] = 'DI'
di_collection <- rbind(acs_di[, c(1, 4, 3)], cs_cogs2, szsafd_cogs2)
di_collection <- di_collection[order(di_collection$City), ]
#View(di_collection)

# barplot of the dataframe
di_bars <- ggplot(data = di_collection, aes(x = City, y = DI, fill = Diagnosis)) + geom_bar(stat = 'identity', position = position_dodge()) + ylab('Diversity Index') + ggtitle('Diversity Index COGS2 Cities') 

di_bars
```

## Hypothesis Testing: Monte-Carlo Simulation

### Getting sample sizes
```{r}
# by-city and diagnosis sample sizes
city_diag_sam_sizes <- plyr::count(no_sd, c('cLocationCity', 'cDiagnosis4'))
city_diag_sam_sizes

# by-diagnosis sample size (aggregate)
city_diag_agg_sam_size <-  plyr::count(no_sd, 'cDiagnosis4')
city_diag_agg_sam_size
```


### Simulation Methodology

Methodology:
1) Follow same for loop structure as in generation of diversity index
2) After generating the proportions vector for use in the DI calculation:
- Randomly sample n from the rows of the weighted_cts data frame
- Choose n based on what COGS2's counts for CS or SZSAFD within the city of interest
- Then re-generate diversity index
3) Store results in a matrix
- 1 row per city per diagnosis
- 10 rows total (include the aggregate)
- 1000 columns

### Generating the Results Matrix
```{r}
city_diag_sam_sizes
city_diag_agg_sam_size

agg_info <- data.frame(cLocationCity = 'Aggregate',
                       cDiagnosis4 = c('CS', 'SZSAFD'),
                       freq = city_diag_agg_sam_size$freq)
city_ns <- rbind(city_diag_sam_sizes, agg_info)
city_ns

city_ns

N_sim <- 1000
sim_mat <- matrix(0, nrow = 10, ncol = N_sim)

# Same for-loop structure as before
n_track <- 1

for (i in 1:length(cogs_cities_acs2)) {
  for (d in 1:length(diagnoses)) {
    
    if (cogs_cities_acs2[i] != -1) {
      df_sub <- acs1865[acs1865$CITY == cogs_cities_acs2[i], ]
       # total effective population for that city
      # need to use this cuz subset by age
      tot <- sum(df_sub$PERWT)
      
      weighted_cts <- plyr::count(df_sub, c('Race2', 'Hispan2', 'SEX'), wt_var = 'PERWT')
      
      props <- weighted_cts$freq / tot
      
      # begin random sampling
      
      for (N in 1:N_sim) {
        row_samples <- sample(1:nrow(weighted_cts), size = city_ns$freq[i], p = props, replace = TRUE)
        
        # generate proportions for the 24 groups
        row_sam_cts <- plyr::count(row_samples)
        prop_for_DI <- row_sam_cts$freq/sum(row_sam_cts$freq)
        
        # generate diversity index
        DI_sam <- mult_ent(prop_for_DI)
        
        # store in matrix
        sim_mat[n_track, N] <- DI_sam
      }
    } else {
      df_sub <- acs1865
      
      # Get counts of each combination, but per city
      weighted_cts <- plyr::count(df_sub, c('Race2', 'Hispan2', 'SEX','CITY'), wt_var = 'PERWT')
      
      # Generate within-city totals
      tots <- plyr::count(df_sub, 'CITY', wt_var = 'PERWT')
      
      # Generate proportions of each combination within each city
      weighted_cts$props <- rep(0, nrow(weighted_cts))
      
      for (r in 1:nrow(weighted_cts)) {
        weighted_cts$props[r] <- weighted_cts$freq[r] / tots$freq[tots$CITY == weighted_cts$CITY[r]]
      }
      
       for (N in 1:N_sim) {
         
         # splitting the weighted_cts into 1 df for each city
         df_LA <- weighted_cts[weighted_cts$CITY == 3730,]
         df_NY <- weighted_cts[weighted_cts$CITY == 4610,]
         df_Ph <- weighted_cts[weighted_cts$CITY == 5330,]
         df_Se <- weighted_cts[weighted_cts$CITY == 6430,]
         
         rownames(df_LA) <- NULL
         rownames(df_NY) <- NULL
         rownames(df_Ph) <- NULL
         rownames(df_Se) <- NULL
         
         # sample sizes
         ss <- city_ns$freq[city_ns$cDiagnosis4 == diagnoses[d]][1:4]
         
        row_samples <- c(sample(1:nrow(df_LA), size = ss[1], p = df_LA$props, replace = TRUE),
                         sample(1:nrow(df_NY), size = ss[2], p = df_NY$props, replace = TRUE),
                         sample(1:nrow(df_Ph), size = ss[3], p = df_Ph$props, replace = TRUE),
                         sample(1:nrow(df_Se), size = ss[4], p = df_Se$props, replace = TRUE))
        
        # generate proportions for the 24 groups
        row_sam_cts <- plyr::count(row_samples)
        prop_for_DI <- row_sam_cts$freq/sum(row_sam_cts$freq)
        
        # generate diversity index
        DI_sam <- mult_ent(prop_for_DI)
        
        # store in matrix
        sim_mat[n_track, N] <- DI_sam
      }
    }
    
   
    n_track <- n_track + 1
  }
}

acs_di$Percent_Max_DI <- acs_di$DI / log(24)
sim_res <- data.frame( city = rep(cogs_cities_acs2, rep(2, 5)), sim_means  = apply(sim_mat, 1, mean))
sim_res
city_ns

acs_di
```

### Calculating the 95% CI
```{r}
CI_df <- as.data.frame(t(apply(sim_mat, MARGIN = 1, FUN = quantile, prob = c(0.025, 0.50, 0.975), simplify = TRUE)))
CI_df

di_cogs2 <- di_collection[di_collection$Diagnosis!='Census ACS',]
rownames(di_cogs2) <- NULL
di_sig <- cbind(di_cogs2, CI_df)
di_sig$Significant <- di_sig$DI < di_sig$`2.5%` | di_sig$DI > di_sig$`97.5%`
di_sig$Sig_Code <- ifelse(di_sig$Significant, '**', '-')
#View(di_sig)

knitr::kable(di_sig)
grid.newpage()
grid.table(di_sig[, -7], rows = NULL)
View(di_sig)
```

### Barplot of results
```{r}
library(ggsignif)
census_medians <- di_sig[, c('City', 'Diagnosis', '50%')]
census_medians$Diagnosis <- paste(census_medians$Diagnosis, 'ACS')
census_medians
head(census_medians)
colnames(census_medians)[3] <- 'DI'

head(di_sig)

di_sig_plot <- rbind(di_sig[, 1:3], census_medians)
di_sig_plot
hyp_test <- ggplot(data = di_sig_plot, aes(City, DI)) + geom_bar(aes(fill = Diagnosis), width = 0.7, stat = 'identity', position = position_dodge(), color = 'black') + 
  ylim(0, 3.0) +
  ylab('Diversity Index') +
  ggtitle('Diversity Index COGS2 Cities') +
  theme_minimal() + theme(panel.grid.major.x = element_line(color = 'darkred'))
  #                         panel.grid.minor.x = element_line(color = 'grey68'))

cslabel.df <- data.frame(City = 0.8 + 0:4,
                       DI = rep(2.5, 5))

szsafdlabel.df <- data.frame(City = 1.2 + 0:4,
                             DI = rep(2.75, 5))

my_colors <- c('darkgreen', 'greenyellow', 'dodgerblue2', 'cyan2')

# hyp_test + scale_fill_brewer(palette="Blues") + geom_text(data = cslabel.df, label = c('**', '-', '**', '**', '-')) + 
#   geom_text(data= szsafdlabel.df, label = c('**', '**', '**', '**', '**'))
hyp_test + scale_fill_manual(values = my_colors) + geom_text(data = cslabel.df, label = c('**', '**', '**', '-', '-')) + 
  geom_text(data= szsafdlabel.df, label = c('**', '**', '**', '**', '-')) + labs(subtitle = " ' ** ' = p < 0.05, ' - ' = NS")

write.csv(di_sig, file = 'div_idx_resam_test.csv', row.names = FALSE)
write.csv(acs_di, file = 'ACS_DI.csv', row.names = FALSE)

#dev.off()
```
